fix: revert test category refactoring
Reverts moving the category interfaces of tests to their own package.
This is because we export the tests as a jar to maven and it could break
application code.
