+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/ParallelIntegrationTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Google LLC
+ * Copyright 2020 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.spanner.categories;
+package com.google.cloud.spanner;
 
-/** Serial Integration Test interface. */
-public interface SerialIntegrationTest {}
+/** Parallel Integration Test interface. */
+public interface ParallelIntegrationTest extends IntegrationTest {}
