+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/SerialIntegrationTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.spanner.categories;
+package com.google.cloud.spanner;
 
-/** Integration Test interface. */
-public interface IntegrationTest {}
+/** Serial Integration Test interface. */
+public interface SerialIntegrationTest {}
