+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/TracerTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,14 @@
  * limitations under the License.
  */
 
-package com.google.cloud.spanner.categories;
+package com.google.cloud.spanner;
 
-/** Parallel Integration Test interface. */
-public interface ParallelIntegrationTest extends IntegrationTest {}
+/**
+ * Tests marked with this {@link org.junit.experimental.categories.Category} will be executed in a
+ * separate execution with the maven-surefire plugin. The tests will be excluded from execution with
+ * the maven-failsafe plugin.
+ *
+ * <p>Separate execution prevents the injection of any custom tracing configuration from interfering
+ * with other tests, as most tracing configuration is stored in static final variables.
+ */
+public interface TracerTest {}
