+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/it/ITInstanceAdminTest.java
@@ -27,7 +27,7 @@ import com.google.cloud.spanner.InstanceConfig;
 import com.google.cloud.spanner.InstanceInfo;
 import com.google.cloud.spanner.IntegrationTestEnv;
 import com.google.cloud.spanner.Options;
-import com.google.cloud.spanner.categories.SerialIntegrationTest;
+import com.google.cloud.spanner.SerialIntegrationTest;
 import com.google.common.collect.Iterators;
 import com.google.spanner.admin.instance.v1.UpdateInstanceMetadata;
 import java.util.ArrayList;
