+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/it/ITPitrCreateDatabaseTest.java
@@ -27,8 +27,8 @@ import com.google.cloud.spanner.DatabaseId;
 import com.google.cloud.spanner.DatabaseNotFoundException;
 import com.google.cloud.spanner.ErrorCode;
 import com.google.cloud.spanner.IntegrationTestEnv;
+import com.google.cloud.spanner.ParallelIntegrationTest;
 import com.google.cloud.spanner.SpannerException;
-import com.google.cloud.spanner.categories.ParallelIntegrationTest;
 import com.google.cloud.spanner.testing.RemoteSpannerHelper;
 import java.util.ArrayList;
 import java.util.Collections;
