+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/it/ITResultSetGetValue.java
@@ -29,13 +29,13 @@ import com.google.cloud.spanner.Database;
 import com.google.cloud.spanner.DatabaseClient;
 import com.google.cloud.spanner.IntegrationTestEnv;
 import com.google.cloud.spanner.Mutation;
+import com.google.cloud.spanner.ParallelIntegrationTest;
 import com.google.cloud.spanner.ResultSet;
 import com.google.cloud.spanner.Statement;
 import com.google.cloud.spanner.Struct;
 import com.google.cloud.spanner.Type;
 import com.google.cloud.spanner.Type.StructField;
 import com.google.cloud.spanner.Value;
-import com.google.cloud.spanner.categories.ParallelIntegrationTest;
 import com.google.cloud.spanner.testing.RemoteSpannerHelper;
 import com.google.common.primitives.Doubles;
 import java.math.BigDecimal;
